+++ gpaw/lcao/scissors.py
@@ -1,15 +1,38 @@
+"""Scissors operator for LCAO."""
+from typing import Sequence, Tuple
+
 import numpy as np
 from ase.units import Ha
+
+from gpaw.hamiltonian import Hamiltonian
+from gpaw.hints import Array2D, Array3D
+from gpaw.kpoint import KPoint
+from gpaw.wavefunctions.base import WaveFunctions
 
 from .eigensolver import DirectLCAO
 
 
 class Scissors(DirectLCAO):
-    def __init__(self, components):
+    def __init__(self, shifts: Sequence[Tuple[float, float, int]]):
+        """Scissors-operator eigensolver.
+
+        The *shifts* are given as a sequence of tuples::
+
+            [(<shift for occupied states>,
+              <shift for unoccupied states>,
+              <number of atoms>),
+             ...]
+
+        Here we open a gap for states on atoms with indices 3, 4 and 5:
+
+        >>> eigensolver = Scissors([(0.0, 0.0, 3),
+        ...                         (-0.5, 0.5, 3)])
+
+        """
         DirectLCAO.__init__(self)
-        self.components = []
-        for homo, lumo, natoms in components:
-            self.components.append((homo / Ha, lumo / Ha, natoms))
+        self.shifts = []
+        for homo, lumo, natoms in shifts:
+            self.shifts.append((homo / Ha, lumo / Ha, natoms))
 
     def write(self, writer):
         writer.write(name='lcao')
@@ -18,7 +41,7 @@ class Scissors(DirectLCAO):
         txt = DirectLCAO.__repr__(self)
         txt += '\n    Scissors operators:\n'
         a1 = 0
-        for homo, lumo, natoms in self.components:
+        for homo, lumo, natoms in self.shifts:
             a2 = a1 + natoms
             txt += (f'      Atoms {a1}-{a2 - 1}: '
                     f'VB: {homo * Ha:+.3f} eV, '
@@ -26,8 +49,14 @@ class Scissors(DirectLCAO):
             a1 = a2
         return txt
 
-    def calculate_hamiltonian_matrix(self, ham, wfs, kpt, Vt_xMM=None,
-                                     root=-1, add_kinetic=True):
+    def calculate_hamiltonian_matrix(self,
+                                     ham: Hamiltonian,
+                                     wfs: WaveFunctions,
+                                     kpt: KPoint,
+                                     Vt_xMM: Array3D = None,
+                                     root: int = -1,
+                                     add_kinetic: bool = True) -> Array2D:
+        """Add scissors operator."""
         H_MM = DirectLCAO.calculate_hamiltonian_matrix(
             self, ham, wfs, kpt, Vt_xMM, root, add_kinetic)
         if kpt.C_nM is None:
@@ -46,7 +75,7 @@ class Scissors(DirectLCAO):
 
         M1 = 0
         a1 = 0
-        for homo, lumo, natoms in self.components:
+        for homo, lumo, natoms in self.shifts:
             a2 = a1 + natoms
             M2 = M1 + sum(setup.nao for setup in ham.setups[a1:a2])